+++ internal/atomicfile/replace_windows.go
@@ -1,0 +1,25 @@
+//go:build windows
+
+package atomicfile
+
+import "golang.org/x/sys/windows"
+
+func replaceFile(source, destination string) error {
+	from, err := windows.UTF16PtrFromString(source)
+	if err != nil {
+		return err
+	}
+	to, err := windows.UTF16PtrFromString(destination)
+	if err != nil {
+		return err
+	}
+	return windows.MoveFileEx(
+		from,
+		to,
+		windows.MOVEFILE_REPLACE_EXISTING|windows.MOVEFILE_WRITE_THROUGH,
+	)
+}
+
+func syncDirectory(string) error {
+	return nil
+}